* Increase or decrease function call from actions will now also try to decay the item. * Extended reload command with weapons and quests. * Corrected TEXTCOLOR_DARKRED value. * Corrected ids of field items and corpses. * Most damage killer and unjustified (bool) is now also accessible in onDeath event. * Attempt to make the NPC shop system work with fluid containers. * Fixed a crash related to trading. * Changed the house system to use rent instead of x gold per house tile. * Corrected some errors in Item::getDescription. * items.xml: allowpickupable, showattributes, added tea and mead to fluidsources. * Reload items will now also reload movements and weapons otherwise it breaks them. * Extended doCreatureSay with three more optional parameters (..., ghost, cid, pos) * New Lua functions: doCreateNpc(name, pos), getItemWeightByUID(uid). * Events from a Lua interface will now be stopped at reload. * Third parameter in doTeleportThing (pushMovement) will now default to false. * doTeleportThing and normal teleports will also update the look direction of a player. * doRelocate will no longer move doors. * ITEM_TYPE_DEPOT is now supported in getTileItemByType. * getItemDescriptions will now also include description in the table it returns. * doTransformItem will no longer try to transform an item to the same ID. * Applied a patch from 0.3 that fixes a bug which would prevent monsters from dropping empty containers (such as key rings). * Corrected condition tick intervals in monsters. * Added support for freezecondition, cursecondition and dazzlecondition in monsters. * Added drownPercent, lifedrainPercent and physicalPercent to monster immunity elements, and also error handling for incorrect elements. * Fixed crashes and compatibility issues in the NPC code. * Corrected RET_BOTHHANDSNEEDTOBEFREE, RET_NOTENOUGHCAPACITY and RET_CANNOTPICKUP cancel messages. * Corrected the "You have been idle for x minutes" message. * Increased knownCreatureList max size to 250 which is supported in newer clients. * Gamemasters can now walk and teleport through creatures. * Fixed some issues with raids. * Removed the blacklist code as it is no longer in use. * The msgcontains function has been improved. * Heavily updated items.xml file. * Items will no longer spawn with item count 0 from the map editor, this should prevent a possible infinite loop/crash with Tile::_getIndexOfThing. * Bumped version to 0.2.7. * Code cleanup, formating & optimizations.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@3557 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ movement.cpp
@@ -737,7 +737,7 @@ uint32_t MoveEvent::EquipItem(MoveEvent* moveEvent, Player* player, Item* item, 
 
 	if(!player->hasFlag(PlayerFlag_IgnoreWeaponCheck) && moveEvent->getWieldInfo() != 0)
 	{
-		if(player->getLevel() < (uint32_t)moveEvent->getReqLevel() || player->getMagicLevel() < (uint32_t)moveEvent->getReqMagLv())
+		if(player->getLevel() < (uint32_t)moveEvent->getReqLevel() || player->getMagicLevel() < moveEvent->getReqMagLv())
 			return 0;
 
 		if(moveEvent->isPremium() && !player->isPremium())
